update install script

+++ install.sh
@@ -1,6 +1,15 @@
 #!/bin/bash -xe
 
-GUI_DIR=$(pwd)
+TOP_DIR=$(pwd)
+PYTHON_VERSION=`python3 -c "import sys; print(f'python{sys.version_info.major}.{sys.version_info.minor}')"`
+# paths where detectorbank-gui stuff will be installed
+LOCAL_INSTALL_DIR="/home/$USER/.local/lib/$PYTHON_VERSION/site-packages"
+LOCAL_BIN="/home/$USER/.local/bin"
+LOCAL_SHARE_APPS="/home/$USER/.local/share/applications"
+# ensure these paths all exist
+mkdir -p $LOCAL_INSTALL_DIR
+mkdir -p $LOCAL_BIN
+mkdir -p $LOCAL_SHARE_APPS
 
 if [ -f /etc/os-release ]; then
   . /etc/os-release
@@ -19,27 +28,43 @@ else
 fi
 
 if [[ "$1" != "--no-install-detectorbank" ]]; then
-
-  # detectorbank dependencies
-  if [[ $OS == "debian" ]]; then
-    sudo apt install python3-dev python3-numpy swig build-essential autoconf-archive pkg-config libtool libcereal-dev librapidxml-dev libfftw3-dev doxygen python3 graphviz python3-tap
-  else
-    sudo dnf install make automake gcc gcc-c++ kernel-devel python3-devel python3-numpy swig autoconf-archive libtool fftw-devel rapidxml-devel cereal-devel doxygen python3 graphviz
+  INSTALL_DETBANK=true
+  
+  # check if DetectorBank already exists
+  if [[ -d "../DetectorBank" ]]; then
+    read -p "../DetectorBank already exists. Do you want to remove it and re-install? [Y/n/c] " answer
+      if [[ -z $answer ]] || [[ $answer != "${answer#[Yy]}" ]]; then 
+        rm -rf "../DetectorBank"
+      elif [[ $answer != "${answer#[Nn]}" ]]; then
+        INSTALL_DETBANK=false
+      else
+        # cancel
+        exit 1
+    fi 
   fi
+  
+  if [ "$INSTALL_DETBANK" = true ]; then
+    # install dependencies
+    if [[ $OS == "debian" ]]; then
+      sudo apt install python3-dev python3-numpy swig build-essential autoconf-archive pkg-config libtool libcereal-dev librapidxml-dev libfftw3-dev doxygen python3 graphviz python3-tap
+    else
+      sudo dnf install make automake gcc gcc-c++ kernel-devel python3-devel python3-numpy swig autoconf-archive libtool fftw-devel rapidxml-devel cereal-devel doxygen python3 graphviz
+    fi
 
-  ## install DetectorBank
-  cd ..
-  git clone https://github.com/keziah55/DetectorBank.git
-  cd DetectorBank
-  autoreconf --install --verbose
-  mkdir -p build
-  cd build
-  ../configure
-  make
-  sudo make install
+    ## install DetectorBank
+    cd ..
+    git clone https://github.com/keziah55/DetectorBank.git
+    cd DetectorBank
+    autoreconf --install --verbose
+    mkdir -p build
+    cd build
+    ../configure
+    make
+    sudo make install
 
-  ## back to gui
-  cd $GUI_DIR
+    ## back to gui
+    cd $TOP_DIR
+  fi
 fi
   
 # gui dependencies
@@ -52,41 +77,38 @@ fi
 if [[ $NAME == *"Debian"* ]]; then
   # 'python3 -m pip install' raises an error on debian, so get it manually
   # no new dependencies to install
-  PYTHON_VERSION=`python3 -c "import sys; print(f'python{sys.version_info.major}.{sys.version_info.minor}')"`
-  LOCAL_INSTALL_DIR="/home/$USER/.local/lib/$PYTHON_VERSION/site-packages"
   mkdir -p deps
   cd deps
   if [[ -e "CustomPyQtObjects" ]]; then
     rm -rf "CustomPyQtObjects"
   fi
   if [[ -e "$LOCAL_INSTALL_DIR/customQObjects" ]]; then
-    rm "$LOCAL_INSTALL_DIR/customQObjects"
+    rm -rf "$LOCAL_INSTALL_DIR/customQObjects"
   fi
   git clone https://github.com/keziah55/CustomPyQtObjects.git
-  mkdir -p $LOCAL_INSTALL_DIR
-  ln -s $(pwd)/CustomPyQtObjects/customQObjects $LOCAL_INSTALL_DIR
+  cp -r $(pwd)/CustomPyQtObjects/customQObjects $LOCAL_INSTALL_DIR
   cd ..
 else
   python3 -m pip install git+https://github.com/keziah55/CustomPyQtObjects.git
 fi
 
-## copy script and icon, make .desktop
+# copy detectorbank-gui source to .local/lib
+if [[ -e "$LOCAL_INSTALL_DIR/detectorbankgui" ]]; then
+    rm -rf "$LOCAL_INSTALL_DIR/detectorbankgui"
+fi
+cp -r "$TOP_DIR/detectorbankgui" $LOCAL_INSTALL_DIR
 
 # copy script to ~/.local/bin
-LOCAL_BIN="/home/$USER/.local/bin"
-mkdir -p $LOCAL_BIN
 if [[ :$PATH: != *:"$LOCAL_BIN":* ]]; then
   export PATH=$PATH:$LOCAL_BIN
 fi
 if [[ -f "$LOCAL_BIN/detectorbank-gui" ]]; then
   rm "$LOCAL_BIN/detectorbank-gui"
 fi
-ln -s $(pwd)/detectorbank-gui $LOCAL_BIN
+cp "$TOP_DIR/main.py" $LOCAL_BIN/detectorbank-gui
 
 # make .desktop file
 cp detectorbank-gui.desktop.template detectorbank-gui.desktop
-ICON_PATH="$GUI_DIR/images/icon.png"
+ICON_PATH="$TOP_DIR/images/icon.png"
 sed -i "s|/path/here/images/icon.png|${ICON_PATH}|g" detectorbank-gui.desktop
-LOCAL_SHARE_APPS="/home/$USER/.local/share/applications"
-mkdir -p $LOCAL_SHARE_APPS
-cp detectorbank-gui.desktop $LOCAL_SHARE_APPS
+mv detectorbank-gui.desktop $LOCAL_SHARE_APPS
